% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq.R
\name{physeq}
\alias{physeq}
\title{physeq constructor}
\usage{
physeq(headings, sequences)
}
\arguments{
\item{headings}{Vector of sequence names}

\item{sequences}{Vector of nucleotide/amino acid sequences}
}
\description{
Creates a new physeq object (a list of sequences) from a vector of headings and sequences, which can then be saved as a FASTA or PHYLIP file.
}
\examples{
x <- physeq(c("Seq 1", "Seq 2"), c("RRARRPR", "PKKKRKV"))
print(x)
x$Headings #Get or modify headings
x$Sequences #Get or modify sequences
WriteFasta(x, "Seqs.fa")
}
