% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastaPhylip.R
\name{WritePhylip}
\alias{WritePhylip}
\title{PHYLIP-format file sequence alignment writer}
\usage{
WritePhylip(x, filename, blocks = 5, blocksize = 10)
}
\arguments{
\item{x}{The physeq object to read from}

\item{filename}{The name of the file to create}

\item{blocks}{The number of sequences per block. Default = 5}

\item{blocksize}{The number of characters per sequence per block. Default = 10}
}
\description{
Writes a PHYLIP multiple sequence alignment file from a physeq object. If sequences are not the same length, there will be an error
}
\examples{
x <- ReadFasta("MSA.fa") #Creates the physeq object 'x'
WritePhylip(x, "MSA.phy")
y <- physeq(c("Seq 1", "Seq 2"), c("RRARRPR", "PKKKRKV"))
WritePhylip(y, "Seqs.phy")
}
